Syncs brave-ui and brave-core
